+++ stories/Plugins/component/comtextMenu.tsx
@@ -52,6 +52,8 @@ const ContextMenu = () => {
         handleMenuClick: (target, item) => {
           console.log(target, item);
         },
+        offsetY: 20,
+        offsetX: 20,
       });
       graph = new Graph({
         container: container.current as string | HTMLElement,
@@ -59,29 +61,47 @@ const ContextMenu = () => {
         height: 500,
         plugins: [contextMenu],
         modes: {
-          default: ['drag-canvas', 'zoom-canvas'],
+          default: ['drag-canvas', 'zoom-canvas', 'drag-node'],
         },
+        defaultNode: {
+          type: 'modelRect',
+          size: [182, 32],
+          labelCfg: {
+            offset: 12,
+            style: {
+              fontSize: 12,
+            },
+          },
+          stateIcon: {
+            show: false
+          },
+          preRect: {
+            show: false
+          },
+        }
       });
       graph.data(data);
       graph.render();
-      let width = 500,
-        height = 500;
-      graph.on('canvas:click', (evt) => {
-        width += 100;
-        height += 50;
-        graph.changeSize(width, height);
-      });
-      graph.on('node:click', (e) => {
-        graph.removeItem(e.item);
-      });
-      graph.on('edge:click', (e) => {
-        graph.removeItem(e.item);
-      });
+      // let width = 500,
+      //   height = 500;
+      // graph.on('canvas:click', (evt) => {
+      //   width += 100;
+      //   height += 50;
+      //   graph.changeSize(width, height);
+      // });
+      // graph.on('node:click', (e) => {
+      //   graph.removeItem(e.item);
+      // });
+      // graph.on('edge:click', (e) => {
+      //   graph.removeItem(e.item);
+      // });
     }
   });
   return (
     <div>
-      <div style={{ backgroundColor: '#000', width: '100px', height: '100px' }}></div>
+      <div style={{ marginBottom: '10px', height: '130px' }}>
+        <button id="create">创建节点</button>
+      </div>
       <div style={{ left: '100px' }} ref={container}></div>
     </div>
   );