+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix: optimized drag-canvas with hidden items;
 - fix: state style with 0 value problem, closes: #2039;
 - fix: layout with webworker leads to twice beforelayout, closes: #2052;
+- fix: context menu with sibling doms of graph container leads to position problem, closes: #2053;
 - feat: fisheye allows user to config the trigger of scaling range(r) and magnify factor(d) by scaleRBy and scaleDBy respectively;
 - feat: add the percent text of magnify factor(d) for fisheye and users are allowed to configure it by show showDPercent.